+++ HospitalManagementSystem/src/main/java/org/example/entity/AppointmentOutcomeRecord.java
@@ -46,15 +46,15 @@ public class AppointmentOutcomeRecord {
     }
 
 
-    public void updateMedicationStatusFromInput() { //Update status of medication.csv
+    public void updateMedicationStatusFromInput() { //Update status of Medicine_List.csv
         Scanner scanner = new Scanner(System.in);
-        System.out.print("Enter the index of the medication.csv status to update (0 to " + (medicationStatus.size() - 1) + "): ");
+        System.out.print("Enter the index of the Medicine_List.csv status to update (0 to " + (medicationStatus.size() - 1) + "): ");
 
         int index = scanner.nextInt(); // Read index
         scanner.nextLine(); // Consume the newline
 
         if (index >= 0 && index < medicationStatus.size()) {
-            System.out.print("Enter the new medication.csv status: ");
+            System.out.print("Enter the new Medicine_List.csv status: ");
             String newStatus = scanner.nextLine(); // Read new status
             medicationStatus.set(index, newStatus); // Update the status
             System.out.println("Medication status updated to: " + newStatus);
