+++ HospitalManagementSystem/src/main/java/org/example/entity/Staff.java
@@ -1,0 +1,4 @@
+package org.example.entity;
+
+public class Staff {
+}
