+++ HospitalManagementSystem/src/main/java/org/example/menu/Menu.java
@@ -1,10 +1,5 @@
 package org.example.menu;
 
-import org.example.control.AccountService;
-import org.example.entity.User;
-
-import java.util.Scanner;
-
 public interface Menu {
     void displayMenu();
     void handleChoice(int choice);
