+++ HospitalManagementSystem/src/main/resources/Medicine_List.csv
@@ -1,0 +1,3 @@
+id,name,quantity
+0,paracetamol,1
+1,ibuprofen,2
