+++ HospitalManagementSystem/src/main/resources/Staff_List.csv
@@ -1,0 +1,5 @@
+Staff ID,Name,Role,Gender,Age
+D001,John Smith,Doctor,Male,45
+D002,Emily Clarke,Doctor,Female,38
+P001,Mark Lee,Pharmacist,Male,29
+A001,Sarah Lee,Administrator,Female,40
