+++ HospitalManagementSystem/src/main/resources/doctor_availability/D001_availability.csv
@@ -1,0 +1,3 @@
+TimeSlot,Monday,Tuesday,Wednesday,Thursday,Friday,Saturday,Sunday
+09:00,Available,Available,Available,Available,Available,Available,Available
+10:00,Available,Unavailable,Unavailable,Available,Unavailable,Available,Available
