--- HospitalManagementSystem/src/main/resources/medication.csv
@@ -1,3 +1,0 @@
-id,name
-0,paracetamol
-1,ibuprofen
