--- HospitalManagementSystem/src/main/resources/userdb.csv
@@ -1,2 +1,0 @@
-username,password,role,contact,id
-example,password,patient,88888888,0
